verbs: Move RMA interface into separate file
