+++ prov/verbs/src/fi_verbs.c
@@ -68,55 +68,6 @@ struct fi_provider fi_ibv_prov = {
 #define VERBS_MSG_ORDER (FI_ORDER_RAR | FI_ORDER_RAW | FI_ORDER_RAS | \
 		FI_ORDER_WAW | FI_ORDER_WAS | FI_ORDER_SAW | FI_ORDER_SAS )
 
-#define VERBS_COMP_READ_FLAGS(ep, flags) ((!VERBS_SELECTIVE_COMP(ep) || \
-		(flags & (FI_COMPLETION | FI_TRANSMIT_COMPLETE | FI_DELIVERY_COMPLETE))) ? \
-		IBV_SEND_SIGNALED : 0)
-
-#define VERBS_COMP_READ(ep) VERBS_COMP_READ_FLAGS(ep, ep->info->tx_attr->op_flags)
-
-
-#define fi_ibv_set_sge(sge, buf, len, desc)		\
-	do {						\
-		sge.addr = (uintptr_t)buf;		\
-		sge.length = (uint32_t)len;		\
-		sge.lkey = (uint32_t)(uintptr_t)desc;	\
-	} while (0)
-
-#define fi_ibv_set_sge_iov(sg_list, iov, count, desc, len)		\
-	do {								\
-		int i;							\
-		if (count) {						\
-			sg_list = alloca(sizeof(*sg_list) * count);	\
-			for (i = 0; i < count; i++) {			\
-				fi_ibv_set_sge(sg_list[i],		\
-						iov[i].iov_base,	\
-						iov[i].iov_len,		\
-						desc[i]);		\
-				len += iov[i].iov_len;			\
-			}						\
-		}							\
-	} while (0)
-
-#define fi_ibv_set_sge_inline(sge, buf, len)	\
-	do {					\
-		sge.addr = (uintptr_t)buf;	\
-		sge.length = (uint32_t)len;	\
-	} while (0)
-
-#define fi_ibv_set_sge_iov_inline(sg_list, iov, count, len)		\
-	do {								\
-		int i;							\
-		if (count) {						\
-			sg_list = alloca(sizeof(*sg_list) * count);	\
-			for (i = 0; i < count; i++) {			\
-				fi_ibv_set_sge_inline(sg_list[i],	\
-						iov[i].iov_base,	\
-						iov[i].iov_len);	\
-				len += iov[i].iov_len;			\
-			}						\
-		}							\
-	} while (0)
-
 
 static const char *local_node = "localhost";
 static char def_tx_ctx_size[16] = "384";
@@ -1048,8 +999,8 @@ err1:
 		return -FI_ENODATA;
 }
 
-static ssize_t fi_ibv_send(struct fi_ibv_msg_ep *ep, struct ibv_send_wr *wr, size_t len,
-		int count, void *context)
+ssize_t fi_ibv_send(struct fi_ibv_msg_ep *ep, struct ibv_send_wr *wr, size_t len,
+		    int count, void *context)
 {
 	struct ibv_send_wr *bad_wr;
 	int ret;
@@ -1106,202 +1057,6 @@ ssize_t fi_ibv_send_iov_flags(struct fi_ibv_msg_ep *ep, struct ibv_send_wr *wr,
 	wr->send_flags = VERBS_INJECT_FLAGS(ep, len, flags) | VERBS_COMP_FLAGS(ep, flags);
 
 	return fi_ibv_send(ep, wr, len, count, context);
-}
-
-static ssize_t
-fi_ibv_msg_ep_rma_write(struct fid_ep *ep_fid, const void *buf, size_t len,
-		     void *desc, fi_addr_t dest_addr,
-		     uint64_t addr, uint64_t key, void *context)
-{
-	struct fi_ibv_msg_ep *ep;
-	struct ibv_send_wr wr;
-
-	memset(&wr, 0, sizeof(wr));
-	wr.opcode = IBV_WR_RDMA_WRITE;
-	wr.wr.rdma.remote_addr = addr;
-	wr.wr.rdma.rkey = (uint32_t) key;
-
-	ep = container_of(ep_fid, struct fi_ibv_msg_ep, ep_fid);
-	wr.send_flags = VERBS_INJECT(ep, len) | VERBS_COMP(ep);
-
-	return fi_ibv_send_buf(ep, &wr, buf, len, desc, context);
-}
-
-static ssize_t
-fi_ibv_msg_ep_rma_writev(struct fid_ep *ep_fid, const struct iovec *iov, void **desc,
-		      size_t count, fi_addr_t dest_addr,
-		      uint64_t addr, uint64_t key, void *context)
-{
-	struct fi_ibv_msg_ep *ep;
-	struct ibv_send_wr wr;
-
-
-	memset(&wr, 0, sizeof(wr));
-	wr.opcode = IBV_WR_RDMA_WRITE;
-	wr.wr.rdma.remote_addr = addr;
-	wr.wr.rdma.rkey = (uint32_t) key;
-
-	ep = container_of(ep_fid, struct fi_ibv_msg_ep, ep_fid);
-	return fi_ibv_send_iov(ep, &wr, iov, desc, count, context);
-}
-
-static ssize_t
-fi_ibv_msg_ep_rma_writemsg(struct fid_ep *ep_fid, const struct fi_msg_rma *msg,
-			uint64_t flags)
-{
-	struct fi_ibv_msg_ep *ep;
-	struct ibv_send_wr wr;
-
-	memset(&wr, 0, sizeof(wr));
-	if (flags & FI_REMOTE_CQ_DATA) {
-		wr.opcode = IBV_WR_RDMA_WRITE_WITH_IMM;
-		wr.imm_data = htonl((uint32_t)msg->data);
-	} else {
-		wr.opcode = IBV_WR_RDMA_WRITE;
-	}
-
-	wr.wr.rdma.remote_addr = msg->rma_iov->addr;
-	wr.wr.rdma.rkey = (uint32_t) msg->rma_iov->key;
-
-	ep = container_of(ep_fid, struct fi_ibv_msg_ep, ep_fid);
-	return fi_ibv_send_msg(ep, &wr, msg, flags);
-}
-
-static ssize_t
-fi_ibv_msg_ep_rma_read(struct fid_ep *ep_fid, void *buf, size_t len,
-		    void *desc, fi_addr_t src_addr,
-		    uint64_t addr, uint64_t key, void *context)
-{
-	struct fi_ibv_msg_ep *ep;
-	struct ibv_send_wr wr;
-
-	memset(&wr, 0, sizeof(wr));
-	wr.opcode = IBV_WR_RDMA_READ;
-	wr.wr.rdma.remote_addr = addr;
-	wr.wr.rdma.rkey = (uint32_t) key;
-
-	ep = container_of(ep_fid, struct fi_ibv_msg_ep, ep_fid);
-	wr.send_flags = VERBS_COMP_READ(ep);
-
-	return fi_ibv_send_buf(ep, &wr, buf, len, desc, context);
-}
-
-static ssize_t
-fi_ibv_msg_ep_rma_readv(struct fid_ep *ep_fid, const struct iovec *iov, void **desc,
-		     size_t count, fi_addr_t src_addr,
-		     uint64_t addr, uint64_t key, void *context)
-{
-	struct fi_ibv_msg_ep *ep;
-	struct ibv_send_wr wr;
-	size_t len = 0;
-
-	memset(&wr, 0, sizeof(wr));
-	wr.opcode = IBV_WR_RDMA_READ;
-	wr.wr.rdma.remote_addr = addr;
-	wr.wr.rdma.rkey = (uint32_t) key;
-
-	ep = container_of(ep_fid, struct fi_ibv_msg_ep, ep_fid);
-	wr.send_flags = VERBS_COMP_READ(ep);
-
-	fi_ibv_set_sge_iov(wr.sg_list, iov, count, desc, len);
-
-	return fi_ibv_send(ep, &wr, len, count, context);
-}
-
-static ssize_t
-fi_ibv_msg_ep_rma_readmsg(struct fid_ep *ep_fid, const struct fi_msg_rma *msg,
-			uint64_t flags)
-{
-	struct fi_ibv_msg_ep *ep;
-	struct ibv_send_wr wr;
-	size_t len = 0;
-
-	memset(&wr, 0, sizeof(wr));
-	wr.opcode = IBV_WR_RDMA_READ;
-	wr.wr.rdma.remote_addr = msg->rma_iov->addr;
-	wr.wr.rdma.rkey = (uint32_t) msg->rma_iov->key;
-
-	ep = container_of(ep_fid, struct fi_ibv_msg_ep, ep_fid);
-	wr.send_flags = VERBS_COMP_READ_FLAGS(ep, flags);
-
-	fi_ibv_set_sge_iov(wr.sg_list, msg->msg_iov, msg->iov_count, msg->desc,	len);
-
-	return fi_ibv_send(ep, &wr, len, msg->iov_count, msg->context);
-}
-
-static ssize_t
-fi_ibv_msg_ep_rma_writedata(struct fid_ep *ep_fid, const void *buf, size_t len,
-			void *desc, uint64_t data, fi_addr_t dest_addr,
-			uint64_t addr, uint64_t key, void *context)
-{
-	struct fi_ibv_msg_ep *ep;
-	struct ibv_send_wr wr;
-
-	memset(&wr, 0, sizeof(wr));
-	wr.opcode = IBV_WR_RDMA_WRITE_WITH_IMM;
-	wr.imm_data = htonl((uint32_t)data);
-	wr.wr.rdma.remote_addr = addr;
-	wr.wr.rdma.rkey = (uint32_t) key;
-
-	ep = container_of(ep_fid, struct fi_ibv_msg_ep, ep_fid);
-	wr.send_flags = VERBS_INJECT(ep, len) | VERBS_COMP(ep);
-
-	return fi_ibv_send_buf(ep, &wr, buf, len, desc, context);
-}
-
-static ssize_t
-fi_ibv_msg_ep_rma_inject_write(struct fid_ep *ep_fid, const void *buf, size_t len,
-		     fi_addr_t dest_addr, uint64_t addr, uint64_t key)
-{
-	struct fi_ibv_msg_ep *ep;
-	struct ibv_send_wr wr;
-
-	memset(&wr, 0, sizeof(wr));
-	wr.opcode = IBV_WR_RDMA_WRITE;
-	wr.wr.rdma.remote_addr = addr;
-	wr.wr.rdma.rkey = (uint32_t) key;
-	wr.send_flags = IBV_SEND_INLINE;
-
-	ep = container_of(ep_fid, struct fi_ibv_msg_ep, ep_fid);
-
-	return fi_ibv_send_buf_inline(ep, &wr, buf, len);
-}
-
-static ssize_t
-fi_ibv_msg_ep_rma_inject_writedata(struct fid_ep *ep_fid, const void *buf, size_t len,
-			uint64_t data, fi_addr_t dest_addr, uint64_t addr,
-			uint64_t key)
-{
-	struct fi_ibv_msg_ep *ep;
-	struct ibv_send_wr wr;
-
-	memset(&wr, 0, sizeof(wr));
-	wr.opcode = IBV_WR_RDMA_WRITE_WITH_IMM;
-	wr.imm_data = htonl((uint32_t)data);
-	wr.wr.rdma.remote_addr = addr;
-	wr.wr.rdma.rkey = (uint32_t) key;
-	wr.send_flags = IBV_SEND_INLINE;
-
-	ep = container_of(ep_fid, struct fi_ibv_msg_ep, ep_fid);
-	return fi_ibv_send_buf_inline(ep, &wr, buf, len);
-}
-
-static struct fi_ops_rma fi_ibv_msg_ep_rma_ops = {
-	.size = sizeof(struct fi_ops_rma),
-	.read = fi_ibv_msg_ep_rma_read,
-	.readv = fi_ibv_msg_ep_rma_readv,
-	.readmsg = fi_ibv_msg_ep_rma_readmsg,
-	.write = fi_ibv_msg_ep_rma_write,
-	.writev = fi_ibv_msg_ep_rma_writev,
-	.writemsg = fi_ibv_msg_ep_rma_writemsg,
-	.inject = fi_ibv_msg_ep_rma_inject_write,
-	.writedata = fi_ibv_msg_ep_rma_writedata,
-	.injectdata = fi_ibv_msg_ep_rma_inject_writedata,
-};
-
-struct fi_ops_rma *fi_ibv_msg_ep_ops_rma(struct fi_ibv_msg_ep *ep)
-{
-	return &fi_ibv_msg_ep_rma_ops;
 }
 
 static int fi_ibv_mr_close(fid_t fid)
