+++ prov/verbs/src/fi_verbs.h
@@ -173,21 +173,6 @@ int fi_ibv_create_ep(const char *node, const char *service,
 		     uint64_t flags, const struct fi_info *hints,
 		     struct rdma_addrinfo **rai, struct rdma_cm_id **id);
 
-ssize_t fi_ibv_send_buf(struct fi_ibv_msg_ep *ep, struct ibv_send_wr *wr,
-			const void *buf, size_t len, void *desc, void *context);
-ssize_t fi_ibv_send_buf_inline(struct fi_ibv_msg_ep *ep, struct ibv_send_wr *wr,
-			       const void *buf, size_t len);
-ssize_t fi_ibv_send_iov_flags(struct fi_ibv_msg_ep *ep, struct ibv_send_wr *wr,
-			      const struct iovec *iov, void **desc, int count,
-			      void *context, uint64_t flags);
-#define fi_ibv_send_iov(ep, wr, iov, desc, count, context)	\
-	fi_ibv_send_iov_flags(ep, wr, iov, desc, count, context,\
-			ep->info->tx_attr->op_flags)
-#define fi_ibv_send_msg(ep, wr, msg, flags)					\
-	fi_ibv_send_iov_flags(ep, wr, msg->msg_iov, msg->desc, msg->iov_count,	\
-			msg->context, flags)
-
-
 struct fi_ops_atomic *fi_ibv_msg_ep_ops_atomic(struct fi_ibv_msg_ep *ep);
 struct fi_ops_cm *fi_ibv_msg_ep_ops_cm(struct fi_ibv_msg_ep *ep);
 struct fi_ops_msg *fi_ibv_msg_ep_ops_msg(struct fi_ibv_msg_ep *ep);
@@ -216,3 +201,64 @@ int fi_ibv_check_rx_attr(const struct fi_rx_attr *attr,
 			 const struct fi_info *hints, const struct fi_info *info);
 int fi_ibv_check_tx_attr(const struct fi_tx_attr *attr,
 			 const struct fi_info *hints, const struct fi_info *info);
+
+
+ssize_t fi_ibv_send(struct fi_ibv_msg_ep *ep, struct ibv_send_wr *wr, size_t len,
+		    int count, void *context);
+ssize_t fi_ibv_send_buf(struct fi_ibv_msg_ep *ep, struct ibv_send_wr *wr,
+			const void *buf, size_t len, void *desc, void *context);
+ssize_t fi_ibv_send_buf_inline(struct fi_ibv_msg_ep *ep, struct ibv_send_wr *wr,
+			       const void *buf, size_t len);
+ssize_t fi_ibv_send_iov_flags(struct fi_ibv_msg_ep *ep, struct ibv_send_wr *wr,
+			      const struct iovec *iov, void **desc, int count,
+			      void *context, uint64_t flags);
+
+#define fi_ibv_set_sge(sge, buf, len, desc)				\
+	do {								\
+		sge.addr = (uintptr_t)buf;				\
+		sge.length = (uint32_t)len;				\
+		sge.lkey = (uint32_t)(uintptr_t)desc;			\
+	} while (0)
+
+#define fi_ibv_set_sge_iov(sg_list, iov, count, desc, len)		\
+	do {								\
+		int i;							\
+		if (count) {						\
+			sg_list = alloca(sizeof(*sg_list) * count);	\
+			for (i = 0; i < count; i++) {			\
+				fi_ibv_set_sge(sg_list[i],		\
+						iov[i].iov_base,	\
+						iov[i].iov_len,		\
+						desc[i]);		\
+				len += iov[i].iov_len;			\
+			}						\
+		}							\
+	} while (0)
+
+#define fi_ibv_set_sge_inline(sge, buf, len)				\
+	do {								\
+		sge.addr = (uintptr_t)buf;				\
+		sge.length = (uint32_t)len;				\
+	} while (0)
+
+#define fi_ibv_set_sge_iov_inline(sg_list, iov, count, len)		\
+	do {								\
+		int i;							\
+		if (count) {						\
+			sg_list = alloca(sizeof(*sg_list) * count);	\
+			for (i = 0; i < count; i++) {			\
+				fi_ibv_set_sge_inline(sg_list[i],	\
+						iov[i].iov_base,	\
+						iov[i].iov_len);	\
+				len += iov[i].iov_len;			\
+			}						\
+		}							\
+	} while (0)
+
+#define fi_ibv_send_iov(ep, wr, iov, desc, count, context)		\
+	fi_ibv_send_iov_flags(ep, wr, iov, desc, count, context,	\
+			ep->info->tx_attr->op_flags)
+
+#define fi_ibv_send_msg(ep, wr, msg, flags)				\
+	fi_ibv_send_iov_flags(ep, wr, msg->msg_iov, msg->desc,		\
+			msg->iov_count,	msg->context, flags)
